tests: c_lib: toupper and tolower take int argument
Typecast the arguments to both toupper() and tolower() to int.
Some ancient newlib has a way to generate warning with -Wall
so we need to be strict about the type of argument.

+++ tests/lib/c_lib/src/main.c
@@ -1011,8 +1011,8 @@ ZTEST(test_c_lib, test_tolower_toupper)
 	int i = 0;
 
 	for (; i < strlen(test); i++) {
-		up[i] = toupper(test[i]);
-		lw[i] = tolower(test[i]);
+		up[i] = toupper((int)test[i]);
+		lw[i] = tolower((int)test[i]);
 	}
 	lw[i] = up[i] = '\0';
